move transferType.go to cliutils
